+++ pkg/cmd/cookie-banner/view/view.go
@@ -48,6 +48,7 @@ query($id: ID!) {
       }
       defaultLanguage
       publisherCountryCode
+      tcfCmpId
       createdAt
       updatedAt
     }
@@ -71,6 +72,7 @@ type viewResponse struct {
 		} `json:"capabilities"`
 		DefaultLanguage      string `json:"defaultLanguage"`
 		PublisherCountryCode string `json:"publisherCountryCode"`
+		TcfCmpId             int    `json:"tcfCmpId"`
 		CreatedAt            string `json:"createdAt"`
 		UpdatedAt            string `json:"updatedAt"`
 	} `json:"node"`
@@ -137,6 +139,7 @@ func NewCmdView(f *cmdutil.Factory) *cobra.Command {
 			_, _ = fmt.Fprintf(out, "%s%d days\n", label.Render("Consent Expiry:"), v.ConsentExpiryDays)
 			_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("Default Language:"), v.DefaultLanguage)
 			_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("Publisher Country:"), v.PublisherCountryCode)
+			_, _ = fmt.Fprintf(out, "%s%d\n", label.Render("TCF CMP ID:"), v.TcfCmpId)
 			_, _ = fmt.Fprintf(out, "%s%t\n", label.Render("Resource Reporting:"), v.Capabilities.ResourceReporting)
 
 			_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("Cookie Policy:"), v.CookiePolicyUrl)